Fix Linux -Wconversion

+++ daemon/shmem_sysv.cpp
@@ -107,7 +107,8 @@ auto create_sysv_shmem_id(int key, std::size_t size, bool force = false,
         if (use_huge_pages) {
             int ret = SHM_HUGETLB;
             if (huge_page_size > 0)
-                ret |= (log2_size(huge_page_size) << SHM_HUGE_SHIFT);
+                ret |= (static_cast<int>(log2_size(huge_page_size))
+                        << SHM_HUGE_SHIFT);
             return ret;
         }
         return 0;
